+++ lib/enums/request_type.dart
@@ -1,3 +1,0 @@
-// Created By Amit Jangid on 14/01/22
-
-enum RequestType { get, post, put, delete }